+++ tt-media-server/tests/test_fine_tuning_api.py
@@ -9,7 +9,7 @@ from fastapi import FastAPI
 from fastapi.testclient import TestClient
 from open_ai_api.fine_tuning import router
 from resolver.service_resolver import service_resolver
-from open_ai_api.fine_tuning import get_org_id
+from security.org_id_checker import get_org_id
 from security.api_key_checker import get_api_key
